refactor(viewer): extract viewer_standings.jsx from viewer.jsx
Pure file split, no behavior change. Step 5 of mp-pxxc.
