+++ web-mobile/index.html
@@ -87,6 +87,10 @@
          mymatchQueueLabel, subBoutLabel), extracted from viewer.jsx (mp-pxxc step 4).
          Must load after viewer_utils.js (its dep) and before viewer.js. -->
     <script type="module" src="/dist/viewer_match.js?v=5"></script>
+    <!-- Pool/standings components (PoolsViewer, LeagueMatrix, SwissStandingsViewer, …),
+         extracted from viewer.jsx (mp-pxxc step 5).
+         Must load after viewer_match.js and before viewer.js. -->
+    <script type="module" src="/dist/viewer_standings.js?v=5"></script>
     <script type="module" src="/dist/viewer.js?v=5"></script>
     <script type="module" src="/dist/admin_helpers.js?v=5"></script>
     <script type="module" src="/dist/admin_shell.js?v=5"></script>
